Added schematic
